First major reorganization step: pulled out the "find sentinels" part of FindSentinels, and made sure all current tests were passing.
